+++ mobile-app/app/_layout.tsx
@@ -9,14 +9,9 @@ import * as SplashScreen from "expo-splash-screen";
 import { StatusBar } from "expo-status-bar";
 import { useEffect, useState } from "react";
 import "react-native-reanimated";
+import SplashScreenUI from "./components/SplashScreen";
 
 import { useColorScheme } from "@/hooks/useColorScheme";
-import { View, Text, StyleSheet } from "react-native";
-import { Colors } from "@/constants/Colors";
-import { LinearGradient } from "expo-linear-gradient";
-import { Fonts } from "@/constants/Fonts";
-import { QueryClientProvider } from "@tanstack/react-query";
-import { queryClient } from "@/queryClient";
 
 // Prevent the splash screen from auto-hiding before asset loading is complete.
 SplashScreen.preventAutoHideAsync();
@@ -34,29 +29,12 @@ export default function RootLayout() {
     SplashScreen.hideAsync();
     setTimeout(() => {
       setSplashVisible(false);
-    }, 3000);
+    }, 5000);
   }, []);
 
   if (splashVisible && !loaded) {
     return (
-      <QueryClientProvider client={queryClient}>
-        <LinearGradient
-          colors={[Colors.light.gradient, Colors.light.primary]}
-          style={styles.container}
-        >
-          <View style={styles.titleContainer}>
-            <Text style={styles.splashTitle}>
-              Lyric
-              <Text style={styles.splashFlip}>Flip</Text>
-            </Text>
-          </View>
-          <View style={styles.bottomContainer}>
-            <Text style={styles.splashSubtitle}>
-              Have fun testing your lyrical knowledge 🎵
-            </Text>
-          </View>
-        </LinearGradient>
-      </QueryClientProvider>
+     <SplashScreenUI />
     );
   }
 
@@ -70,38 +48,3 @@ export default function RootLayout() {
     </ThemeProvider>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: "space-between",
-    alignItems: "center",
-    paddingHorizontal: 20,
-    backgroundColor: Colors.light.gradient,
-  },
-  titleContainer: {
-    flex: 1,
-    justifyContent: "center",
-    alignItems: "center",
-  },
-  splashTitle: {
-    fontSize: 48,
-    color: "white",
-    fontFamily: Fonts.Bricolage,
-  },
-  bottomContainer: {
-    justifyContent: "flex-end",
-    width: "100%",
-    alignItems: "center",
-    paddingBottom: 30,
-  },
-  splashSubtitle: {
-    fontSize: 16,
-    color: "white",
-    fontFamily: Fonts.Inter,
-  },
-  splashFlip: {
-    color: Colors.light.tetiary,
-    fontFamily: Fonts.Bricolage,
-  },
-});
